requirements.txt
Django==3.2.7
django-filter==2.4.0
django-tables2==2.4.0
django-bootstrap3==15.0.0
django-crispy-forms==1.12.0
mysql-connector-python==8.0.26
crispy-bootstrap5==0.5